+++ routes/__init__.py
@@ -2,5 +2,5 @@
 routes package — API routers for Commander AI Lab.
 
 Each module defines a FastAPI APIRouter with real endpoint implementations.
-All shared helpers, models, and state live in routes.shared.
+Shared helpers live in services/, models live in models/.
 """
