+++ routes/deepseek.py
@@ -28,12 +28,10 @@ from fastapi import APIRouter, HTTPException
 from fastapi import Request as FastAPIRequest
 from fastapi.responses import JSONResponse
 
-from routes.shared import (
-    CFG, log_sim,
-    _get_db_conn,
-    _get_deepseek_brain,
-    _ml_logging_enabled,
-)
+from models.state import CFG
+from services.database import _get_db_conn
+from services.forge_runner import _get_deepseek_brain
+from services.logging import log_sim, _ml_logging_enabled
 from commander_ai_lab.sim.validator_brain import ValidatorBrain, ValidatorConfig
 logger = logging.getLogger(__name__)
